requirements-dev.txt
twine
flake8
pylint
safety
pip-audit
pytest
pycryptodome